refactor: Split map prep into leaflet caches and tab prep modules
Extract Leaflet session LRU/export helpers to app_prep_map_leaflet_caches
and non-map tab prep spinner to app_prep_map_tab_prep. Keeps per-mode
GeoJSON builders in app_prep_map_ui for a smaller phase-1 hygiene split.

Refs: #243
